~p + bug

+++ ft_itoa_base.c
@@ -6,13 +6,13 @@
 /*   By: amoinier <amoinier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/23 15:52:46 by amoinier          #+#    #+#             */
-/*   Updated: 2016/03/15 12:33:56 by amoinier         ###   ########.fr       */
+/*   Updated: 2016/03/15 14:15:24 by amoinier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "printf.h"
 
-int		size_nb(unsigned int nb)
+int		size_nb(unsigned long long nb)
 {
 	int	i;
 
@@ -25,16 +25,16 @@ int		size_nb(unsigned int nb)
 	return (i);
 }
 
-void	ft_numb(unsigned int nb, int nbase, char *s, int *i)
+void	ft_numb(unsigned long long nb, int nbase, char *s, int *i)
 {
 	char	const	base[] = "0123456789ABCDEF";
 
-	if (nb >= (unsigned int)nbase)
+	if (nb >= (unsigned long long)nbase)
 		ft_numb(nb / nbase, nbase, s, i);
 	s[(*i)++] = base[nb % nbase];
 }
 
-char	*ft_itoa_base(unsigned int nb, int nbase)
+char	*ft_itoa_base(unsigned long long nb, int nbase)
 {
 	char	*s;
 	int	i;
