+++ printf.h
@@ -6,7 +6,7 @@
 /*   By: amoinier <amoinier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/15 11:11:22 by amoinier          #+#    #+#             */
-/*   Updated: 2016/03/15 11:20:26 by amoinier         ###   ########.fr       */
+/*   Updated: 2016/03/15 14:14:43 by amoinier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,6 @@ int				ft_atoi(char *s);
 size_t			ft_strlen(const char *s);
 char			*ft_strdup(const char *s1);
 
-char			*ft_itoa_base(unsigned int nb, int nbase);
+char			*ft_itoa_base(unsigned long long nb, int nbase);
 
 #endif
